requirements.txt
# This file is autogenerated by pip-compile with the following command:
# pip-compile requirements.in
# To update, run:
# pip-compile requirements.in

click==8.0.3
fastapi==0.70.0
uvicorn==0.15.0
boto3==1.18.66
requests==2.26.0
nemo_toolkit==1.0.0
pdoc==7.0.1
pytest==6.2.5
flake8==4.0.1
python-dotenv==0.19.2
